web-ui/backend/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
python-multipart
pyyaml
pydantic
markdown
python-dotenv
